# perf: ~90% faster `frappe.get_cached_doc`