## Redis-only cache testing

### Before
```py
In [6]: %timeit frappe.get_cached_doc("Sales Invoice", "SINV-22-00016")
215 µs ± 988 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [7]: %timeit frappe.get_cached_doc("System Settings")
34.9 µs ± 83.9 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

### After

```py


In [3]: %timeit frappe.get_cached_doc("Sales Invoice", "SINV-22-00016")
4.15 µs ± 6.44 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each) # 98% improvement

In [4]: %timeit frappe.get_cached_doc("System Settings")
4.14 µs ± 33.4 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each) # 88% improvement


```

Note that `frappe.get_cached_value` is slightly (4%) slower due to unpickling `Document` instead of `_dict`

[ERPNext Run](https://github.com/frappe/erpnext/actions/runs/2460739957)